moved abstract weapon classes to their own package
