+++ app/src/main/java/com/philipgreen/dmwizard/weapons/abstractWeapons/MeleeWeapon.java
@@ -1,4 +1,6 @@
-package com.philipgreen.dmwizard.weapons;
+package com.philipgreen.dmwizard.weapons.abstractWeapons;
+
+import com.philipgreen.dmwizard.weapons.abstractWeapons.BaseWeapon;
 
 /**
  * Created by pgreen on 8/7/16.
